[app] Replace NewProjectCard with NewProjectButton
