+++ app/src/app/dashboard/NewProjectButton.tsx
@@ -1,0 +1,38 @@
+'use client';
+
+import { useState } from 'react';
+import { Plus } from 'lucide-react';
+import { Button } from '#app/components/ui/button.tsx';
+import { createProject } from '#app/actions/project.ts';
+import { ProjectDialog } from './ProjectDialog';
+
+export function NewProjectButton() {
+  const [open, setOpen] = useState(false);
+
+  function handleClose() {
+    setOpen(false);
+  }
+
+  return (
+    <>
+      <Button
+        onClick={() => {
+          setOpen(true);
+        }}
+      >
+        <Plus />
+        New project
+      </Button>
+      <ProjectDialog
+        action={(estimatedDataVolume, collectionPeriod, subreddits, aiMlModelPlan, formData) =>
+          createProject(estimatedDataVolume, collectionPeriod, subreddits, aiMlModelPlan, formData)
+        }
+        infoMessage="This information initialises your DPIA and will flow through all PETLP phases."
+        open={open}
+        onClose={handleClose}
+        submitLabel="Create project"
+        title="Create new research project"
+      />
+    </>
+  );
+}
